+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 2.20.2
 
+- `Fix` — Append default Tunes if user tunes are provided for Block Tool [#1640](https://github.com/codex-team/editor.js/issues/1640)
 - `Fix` - Prevent the leak of codex-tooltip when Editor.js is destroyed [#1475](https://github.com/codex-team/editor.js/issues/1475).
 
 ### 2.20.1